requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
requests
PyGithub
unidiff
faiss-cpu
celery
redis
pytest
bandit
semgrep
sentry-sdk
typer
rich